+++ Universeauto/Models/ICarRepository.cs
@@ -3,7 +3,7 @@
     public interface ICarRepository
     {
         IEnumerable<Car> Cars { get; }
-        Car GetCar(int id);
+        Car GetCar(long id);
         void AddCar(Car car);
         void UpdateCar(Car car);
         void DeleteCar(Car car);
